hookutils: Tcl/Tk: move tkinter-fully-usable check into TclTkInfo class
Implement the check for `tkinter` being fully-functional as a
property of `TclTkInfo` hook utility class. This will allow us
to use the check in tests other than just `test_tkinter_functional`,
and also allows us to cache the result.
